+++ FloppyControlApp.net7/WindowsFormsApplication2/WindowsFormsApplication2/MyClasses/FormHelpers/Models/FloppyControl.Badsectorkeyval.cs
@@ -5,7 +5,7 @@ public partial class FloppyControl
 	public class Badsectorkeyval
         {
             public string Name { get; set; }
-            public int Id { get; set; }
-            public int Threadid { get; set; }
+            public long Id { get; set; }
+            public long Threadid { get; set; }
         }
     }
