+++ FloppyControlApp.net7/WindowsFormsApplication2/WindowsFormsApplication2/MyClasses/Processing/ProcessingTypes/ProcTypeAdaptiveEntropy.cs
@@ -19,10 +19,10 @@ namespace FloppyControlApp.MyClasses.Processing.ProcessingTypes
 
         public AdaptiveEntropyResult ProcTypeAdaptiveEntropy(ProcTypeArgs ProctypeArgs, long ThreadId, ref int Stop)
         {
-            int i;
+			long i;
             int value;
 
-            int MINUS, FOURUS, SIXUS, EIGHTUS, start, end;
+			long MINUS, FOURUS, SIXUS, EIGHTUS, start, end;
             float RateOfChange;
             System.Diagnostics.Stopwatch SW = new System.Diagnostics.Stopwatch();
